'CMakeLists.txt' (LArCalibTools-00-09-60)
	* Tagging LArCalibTools-00-09-60.
	* Cabling tools moved from LArTools to LArCabling.
	* Tagging LArCalibTools-00-09-59.
	* Fix clang warning.
	* Tagging LArCalibTools-00-09-58.
	* Do not use using namespace in a header.
	* Tagging LArCalibTools-00-09-57.
	* Comply with ATLAS naming conventions.

07-12-2015 Pavol Strizenec <pavol @ mail.cern.ch>
	* changes in LArWFParams2Ntuple and LArCommConditions2Ntuple
	* fix in CompletetoFlat for DSPThreshold
	* tag LArCalibTools-00-09-56

2-12-2015 Pavol Strizenec pavol @ mail.cern.ch
	* adding the LArOFCBin algo
	* not tagging yet

2015-08-31 Walter Lampl <walter.lampl@ cern.ch>
...
(Long ChangeLog diff - truncated)

+++ LArCalorimeter/LArCalibTools/LArCalibTools/LArWaves2Ntuple.h
@@ -42,7 +42,7 @@ class LArWaves2Ntuple : public  LArCond2NtupleBase
   NTuple::Array<double> m_time;
   NTuple::Array<double> m_ampliPulse;
   NTuple::Array<double> m_errorPulse;
-  NTuple::Array<long>   m_triggers;
+  NTuple::Array<int>   m_triggers;
 
   // derived information
   NTuple::Item<double>  m_tmaxAmp;
